[Package] Workaround for custom entries, pass through functor arg
The custom entry machinery behaves a bit differently from what we expected when adding symbols such
as 'nat'. The notations in defined `pkg_laws.v` now contain a workaround that should improve
compatibility with standard Coq syntax.

As the files are all parametrised by a module of type `ProbRulesParam`, they pass though their
functor argument to the preceeding modules. In order to avoid ambiguity, each module in turn
includes the preceding one. This solution is not very modular as it creates a linear dependency
chain, and ideally we should refactor the code to make the dependence on `ProbRulesParam` more
local to a few definitions/lemmas.